+++ mycli/packages/special/main.py
@@ -112,6 +112,8 @@ def quit(*_args):
 
 @special_command('\\e', '\\e', 'Edit command with editor (uses $EDITOR).',
                  arg_type=NO_QUERY, case_sensitive=True)
+@special_command('\\clip', '\\clip', 'Copy query to the system clipboard.',
+                 arg_type=NO_QUERY, case_sensitive=True)
 @special_command('\\G', '\\G', 'Display current query results vertically.',
                  arg_type=NO_QUERY, case_sensitive=True)
 def stub():